managin scipionApp

+++ xmipp3/__init__.py
@@ -32,6 +32,8 @@ import pwem
 from pyworkflow import Config
 import pyworkflow.utils as pwutils
 from scipion.install.funcs import CommandDef
+from scipion import __version__ as scipionAppVersion
+from packaging.version import Version
 from .base import *
 from .version import *
 from .constants import XMIPP_HOME, XMIPP_URL, XMIPP_DLTK_NAME, XMIPP_CUDA_BIN, XMIPP_CUDA_LIB, XMIPP_GIT_URL
@@ -167,26 +169,31 @@ class Plugin(pwem.Plugin):
                 default=False
             )
         
-        tag = version._current_xmipp_tag
-        xmippSrc = f'xmippSrc-{tag}'
-        installCommands = [
-            (f'cd .. && rm -rf {xmippSrc} && '
-            f'git clone --depth 1 --branch {tag} {XMIPP_GIT_URL} {xmippSrc} && '
-            f'cd {xmippSrc} && '
-            f'./xmipp -b {tag}', COMPILE_TARGETS)   
-        ]
-        env.addPackage(
-            'xmippSrc', version=tag,
-            tar='void.tgz',
-            commands=installCommands,
-            neededProgs=['git', 'gcc', 'g++', 'cmake', 'make'],
-            updateCuda=True,
-            default=not develMode
-        )
-
-        ## EXTRA PACKAGES ##
-        installDeepLearningToolkit(cls, env)
-
+        #Release binaries installation
+        vScipionApp = Version(scipionAppVersion)
+        vMinRequired = Version('3.7.1')
+        if vScipionApp < vMinRequired:
+	        print("\n---------------------------------------------------\nAttention! The current version of 'scipion-app' is outdated.\nTo update it to the latest version, please run the following command in your terminal:\n\n  scipion3 update\n\nThis command will update 'scipion-app' to the latest available version.\nNote: The minimum required version of 'scipion-app' is 3.7.1.")
+        else:
+            tag = version._current_xmipp_tag
+            xmippSrc = f'xmippSrc-{tag}'
+            installCommands = [
+                (f'cd .. && rm -rf {xmippSrc} && '
+                f'git clone --depth 1 --branch {tag} {XMIPP_GIT_URL} {xmippSrc} && '
+                f'cd {xmippSrc} && '
+                f'./xmipp -b {tag}', COMPILE_TARGETS)   
+            ]
+            env.addPackage(
+                'xmippSrc', version=tag,
+                tar='void.tgz',
+                commands=installCommands,
+                neededProgs=['git', 'gcc', 'g++', 'cmake', 'make'],
+                updateCuda=True,
+                default=not develMode
+            )
+            
+            ## EXTRA PACKAGES ##
+            installDeepLearningToolkit(cls, env)
 
     @classmethod
     def __getBundleDirectory(cls):
